extracting template functions to src
